Move CameraSettings to its own file
