+++ unity/Assets/Maroon/scenes/experiments/OpticsSimulations/Scripts/Camera/CameraSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac97fd222b648864782aa9fe32cdab17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
